+++ src/render.h
@@ -4,12 +4,10 @@
 #include "fractal.h"
 extern Complex C;
 
-// extern float offsets_x[SAMPLE_GRID_WIDTH];
-// extern float offsets_y[SAMPLE_GRID_WIDTH];
-extern float y_max;
-extern float y_min;
-extern float x_max;
-extern float x_min;
+extern double y_max;
+extern double y_min;
+extern double x_max;
+extern double x_min;
 
 void render_fractal(void);
 void initialise_offsets(void);
